Renamed common pkg into core
