+++ internal/core/core_test.go
@@ -1,4 +1,4 @@
-package common
+package core
 
 import (
 	"fmt"
