+++ util.go
@@ -1,15 +1,15 @@
 package rill
 
-import "github.com/destel/rill/internal/common"
+import "github.com/destel/rill/internal/core"
 
 // Drain consumes and discards all items from an input channel, blocking until the channel is closed
 func Drain[A any](in <-chan A) {
-	common.Drain(in)
+	core.Drain(in)
 }
 
 // DrainNB is a non-blocking version of [Drain].
 func DrainNB[A any](in <-chan A) {
-	common.DrainNB(in)
+	core.DrainNB(in)
 }
 
 // Buffer takes a channel of items and returns a buffered channel of exact same items in the same order.
@@ -20,7 +20,7 @@ func DrainNB[A any](in <-chan A) {
 //	ids = Buffer(ids, 100)
 //	// Now up to 100 ids can be buffered if subsequent stages of the pipeline are slow
 func Buffer[A any](in <-chan A, n int) <-chan A {
-	return common.Buffer(in, n)
+	return core.Buffer(in, n)
 }
 
 // FromSlice converts a slice into a channel.
